fixed cleanup script and added new script for manual updates

+++ cleanup-pr-containers.sh
@@ -1,4 +1,6 @@
-containerId=$(docker ps -q | tail -n 1)
+containerId=$(docker ps -q | xargs docker inspect --format="{{.Name}}" $1 | grep -v "hoegg-site" | tail -n 1)
 
-docker stop $containerId
-docker rm $containerId
+if [[ $containerId != "" ]]; then
+	docker stop $containerId
+	docker rm $containerId
+fi
